Added support for 32 bit signal frames

+++ macan/include/macan.h
@@ -54,7 +54,7 @@ void macan_set_ltk(struct macan_ctx *ctx, uint8_t *key);
 void macan_request_keys(struct macan_ctx *ctx, int s);
 int  macan_wait_for_key_acks(struct macan_ctx *ctx, int s);
 int  macan_reg_callback(struct macan_ctx *ctx, uint8_t sig_num, sig_cback fnc);
-void macan_send_sig(struct macan_ctx *ctx, int s, uint8_t sig_num, uint16_t signal);
+void macan_send_sig(struct macan_ctx *ctx, int s, uint8_t sig_num, uint32_t signal);
 int  macan_process_frame(struct macan_ctx *ctx, int s, const struct can_frame *cf);
 
 #endif /* MACAN_H */
